@@ -2,64 +2,74 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.1.3",
+        "php": "^8.1",
         "ext-iconv": "*",
-        "calderacc/geobasic": "^0.2.0",
-        "calderacc/yourls-api-manager": "^0.2.0",
-        "doctrine/doctrine-migrations-bundle": "^1.0",
-        "dragonmantank/cron-expression": "^2.1",
-        "fresh/doctrine-enum-bundle": "^6.2",
-        "friendsofsymfony/elastica-bundle": "^5.0",
-        "friendsofsymfony/jsrouting-bundle": "^2.2",
-        "geocoder-php/nominatim-provider": "^5.0",
-        "hwi/oauth-bundle": "^0.6.1",
-        "jms/serializer-bundle": "^2.3",
-        "jublonet/codebird-php": "^3.1",
-        "league/csv": "^9.1",
-        "league/geotools": "^0.8.0",
-        "maxh/php-nominatim": "^1.2",
-        "nelmio/api-doc-bundle": "^3.1",
-        "php-curl-class/php-curl-class": "^8.0",
-        "php-http/guzzle6-adapter": "^1.1",
-        "php-http/httplug-bundle": "^1.9",
+        "ext-json": "*",
+        "ext-simplexml": "*",
+        "composer-plugin-api": "^2.2",
+        "calderacc/geobasic": "^0.5",
+        "doctrine/annotations": "^1.0",
+        "doctrine/doctrine-bundle": "^2.6",
+        "doctrine/doctrine-migrations-bundle": "^3.1",
+        "doctrine/orm": "^2.10",
+        "flagception/flagception-bundle": "^4.1",
+        "fresh/doctrine-enum-bundle": "^9.0",
+        "friendsofsymfony/elastica-bundle": "^6.2",
+        "friendsofsymfony/jsrouting-bundle": "^3.0",
+        "geocoder-php/nominatim-provider": "^5.2",
+        "jms/serializer-bundle": "^4.0",
+        "jsor/doctrine-postgis": "^2.1",
+        "league/geotools": "^1.0",
+        "maltehuebner/impress-bundle": "^0.2",
+        "mhujer/breadcrumbs-bundle": "^1.5",
+        "nelmio/api-doc-bundle": "^4.0",
+        "nesbot/carbon": "^2.7",
+        "php-amqplib/rabbitmq-bundle": "^2.11",
+        "php-http/guzzle6-adapter": "^2.0.2",
+        "php-http/httplug-bundle": "^1.20.0",
         "php-http/message": "^1.7",
-        "sensio/framework-extra-bundle": "^5.1",
-        "sonata-project/admin-bundle": "^3.40",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.6",
-        "sonata-project/seo-bundle": "^2.5",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "phpstan/phpdoc-parser": "^1.2",
+        "predis/predis": "^1.1",
+        "presta/sitemap-bundle": "^3.0",
+        "sonata-project/seo-bundle": "^3.0",
         "symfony/apache-pack": "^1.0",
-        "symfony/asset": "^4.0",
-        "symfony/console": "^4.0",
-        "symfony/debug-pack": "*",
-        "symfony/expression-language": "^4.0",
-        "symfony/flex": "^1.0",
-        "symfony/form": "^4.0",
-        "symfony/framework-bundle": "^4.0",
-        "symfony/lts": "^4@dev",
-        "symfony/monolog-bundle": "^3.1",
-        "symfony/orm-pack": "*",
-        "symfony/process": "^4.0",
-        "symfony/security-bundle": "^4.0",
-        "symfony/serializer-pack": "*",
-        "symfony/swiftmailer-bundle": "^3.1",
-        "symfony/validator": "^4.0",
-        "symfony/web-link": "^4.0",
-        "symfony/webpack-encore-pack": "*",
-        "symfony/yaml": "^4.0"
+        "symfony/dotenv": "^6.2",
+        "symfony/flex": "^2.0",
+        "symfony/framework-bundle": "^6.2",
+        "symfony/monolog-bundle": "^3.0",
+        "symfony/process": "^6.2",
+        "symfony/property-access": "^6.2",
+        "symfony/proxy-manager-bridge": "^6.2",
+        "symfony/runtime": "^6.2",
+        "symfony/serializer": "^6.2",
+        "symfony/twig-bridge": "^6.2",
+        "symfony/webpack-encore-bundle": "^1.7",
+        "symfony/yaml": "^6.2",
+        "twig/extra-bundle": "^2.12|^3.0",
+        "willdurand/geocoder-bundle": "^5.0"
     },
     "require-dev": {
-        "symfony/browser-kit": "^4.0",
-        "symfony/css-selector": "^4.0",
-        "symfony/dotenv": "^4.0",
+        "phpunit/phpunit": "^9.0",
+        "rector/rector": "^0.14.5",
+        "symfony/browser-kit": "^6.2",
+        "symfony/css-selector": "^6.2",
+        "symfony/debug-bundle": "^6.2",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "^4.0",
-        "symfony/profiler-pack": "*"
+        "symfony/phpunit-bridge": "^4.3",
+        "symfony/web-profiler-bundle": "^6.2"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true,
+            "composer/installers": true,
+            "symfony/flex": true,
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
