@@ -5,37 +5,26 @@ namespace App\Command;
 use App\Entity\City;
 use App\Entity\Station;
 use App\Geocoding\Guesser\CityGuesserInterface;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Symfony\Component\Console\Question\Question;
 
+#[AsCommand(
+    name: 'luft:assign-station',
+    description: 'Assign station to city'
+)]
 class AssignStationCommand extends Command
 {
-    /** @var RegistryInterface $registry */
-    protected $registry;
-
-    /** @var CityGuesserInterface $cityGuesser */
-    protected $cityGuesser;
-
-    public function __construct(?string $name = null, RegistryInterface $registry, CityGuesserInterface $cityGuesser)
+    public function __construct(protected ManagerRegistry $registry, protected CityGuesserInterface $cityGuesser)
     {
-        $this->registry = $registry;
-        $this->cityGuesser = $cityGuesser;
-
-        parent::__construct($name);
+        parent::__construct();
     }
 
-    protected function configure(): void
-    {
-        $this
-            ->setName('luft:assign-station')
-            ->setDescription('Assign station to city');
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): void
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $stationList = $this->registry->getRepository(Station::class)->findWithoutCity();
 
@@ -75,6 +64,9 @@ class AssignStationCommand extends Command
 
             $this->registry->getManager()->flush();
         }
+
+        //return Command::SUCCESS;
+        return 0;
     }
 
     protected function generateCitySlugByCityName(string $cityName): string
