@@ -7,73 +7,40 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as JMS;
 
-/**
- * @ORM\Entity(repositoryClass="App\Repository\CityRepository")
- * @ORM\Table(name="city")
- * @JMS\ExclusionPolicy("ALL")
- */
-class City
+#[ORM\Table(name: 'city')]
+#[ORM\Entity(repositoryClass: 'App\Repository\CityRepository')]
+#[JMS\ExclusionPolicy('ALL')]
+class City implements \Stringable
 {
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
+    #[ORM\Id]
+    #[ORM\Column(type: 'integer')]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
+    protected ?int $id = null;
 
-    /**
-     * @ORM\Column(type="datetime", nullable=false)
-     * @JMS\Expose()
-     * @JMS\Type("DateTime<'U'>")
-     */
-    protected $createdAt;
+    #[ORM\Column(type: 'datetime', nullable: false)]
+    #[JMS\Expose]
+    #[JMS\Type("DateTime<'U'>")]
+    protected ?\DateTime $createdAt = null;
 
-    /**
-     * @ORM\Column(type="string", nullable=false)
-     * @JMS\Expose()
-     */
-    protected $name;
+    #[ORM\Column(type: 'string', nullable: false)]
+    #[JMS\Expose]
+    protected ?string $name = null;
 
-    /**
-     * @ORM\Column(type="string", nullable=false)
-     * @JMS\Expose()
-     */
-    protected $slug;
+    #[ORM\Column(type: 'string', nullable: false)]
+    #[JMS\Expose]
+    protected ?string $slug = null;
 
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     * @JMS\Expose()
-     */
-    protected $description;
+    #[ORM\Column(type: 'string', nullable: true)]
+    #[JMS\Expose]
+    protected ?string $description = null;
 
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     * @JMS\Expose()
-     */
-    protected $fahrverboteSlug;
-
-    /**
-     * @ORM\OneToMany(targetEntity="TwitterSchedule", mappedBy="city")
-     */
-    protected $twitterSchedules;
-
-    /**
-     * @ORM\OneToMany(targetEntity="Station", mappedBy="city")
-     */
-    protected $stations;
-
-    /**
-     * @ORM\OneToOne(targetEntity="User", mappedBy="city")
-     * @ORM\JoinColumn(name="user_id", referencedColumnName="id")
-     */
-    protected $user;
+    #[ORM\OneToMany(targetEntity: 'Station', mappedBy: 'city')]
+    protected Collection $stations;
 
     public function __construct()
     {
         $this->createdAt = new \DateTime();
-        $this->twitterSchedules = new ArrayCollection();
         $this->stations = new ArrayCollection();
-        $this->users = new ArrayCollection();
     }
 
     public function getId(): int
@@ -129,32 +96,6 @@ class City
         return $this;
     }
 
-    public function addTwitterSchedule(TwitterSchedule $twitterSchedule): City
-    {
-        $this->twitterSchedules->add($twitterSchedule);
-
-        return $this;
-    }
-
-    public function getTwitterSchedules(): Collection
-    {
-        return $this->twitterSchedules;
-    }
-
-    public function setTwitterSchedules(Collection $twitterSchedules): City
-    {
-        $this->twitterSchedules = $twitterSchedules;
-
-        return $this;
-    }
-
-    public function removeTwitterSchedule(TwitterSchedule $twitterSchedule): City
-    {
-        $this->twitterSchedules->removeElement($twitterSchedule);
-
-        return $this;
-    }
-
     public function addStation(Station $station): City
     {
         $this->stations->add($station);
@@ -167,9 +108,9 @@ class City
         return $this->stations;
     }
 
-    public function setStations(Collection $twitterSchedules): City
+    public function setStations(Collection $stations): City
     {
-        $this->stations = $twitterSchedules;
+        $this->stations = $stations;
 
         return $this;
     }
@@ -181,37 +122,8 @@ class City
         return $this;
     }
 
-    public function setUser(User $user = null): City
+    public function __toString(): string
     {
-        $this->user = $user;
-
-        return $this;
-    }
-
-    public function getUser(): ?User
-    {
-        return $this->user;
-    }
-
-    public function setFahrverboteSlug(string $fahrverboteSlug): City
-    {
-        $this->fahrverboteSlug = $fahrverboteSlug;
-
-        return $this;
-    }
-
-    public function getFahrverboteSlug(): ?string
-    {
-        return $this->fahrverboteSlug;
-    }
-
-    public function hasFahrverbote(): bool
-    {
-        return $this->fahrverboteSlug !== null;
-    }
-
-    public function __toString(): ?string
-    {
-        return $this->name ? $this->name : '';
+        return $this->name ?: '';
     }
 }
