@@ -5,6 +5,9 @@ namespace App\Entity;
 use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as JMS;
 
+#[ORM\Table(name: 'data')]
+#[ORM\Entity(repositoryClass: 'App\Repository\DataRepository')]
+#[JMS\ExclusionPolicy('ALL')]
 /**
  * @ORM\Entity(repositoryClass="App\Repository\DataRepository")
  * @ORM\Table(name="data", uniqueConstraints={
@@ -14,44 +17,47 @@ use JMS\Serializer\Annotation as JMS;
  */
 class Data
 {
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
+    #[ORM\Id]
+    #[ORM\Column(type: 'integer')]
+    #[ORM\GeneratedValue(strategy: 'AUTO')]
+    protected ?int $id = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity="Station", inversedBy="datas")
-     * @ORM\JoinColumn(name="station_id", referencedColumnName="id")
-     */
-    protected $station;
+    #[ORM\ManyToOne(targetEntity: 'Station', inversedBy: 'datas')]
+    #[ORM\JoinColumn(name: 'station_id', referencedColumnName: 'id')]
+    #[JMS\Expose]
+    #[JMS\Type('App\Entity\Station')]
+    protected ?Station $station = null;
 
-    /**
-     * @ORM\Column(type="datetime", nullable=false)
-     * @JMS\Expose()
-     * @JMS\Type("DateTime<'U'>")
-     */
-    protected $dateTime;
+    #[ORM\Column(type: 'datetime', nullable: false)]
+    #[JMS\Expose]
+    #[JMS\Type("DateTime<'U'>")]
+    protected ?\DateTime $dateTime = null;
 
-    /**
-     * @ORM\Column(type="float", nullable=false)
-     * @JMS\Expose()
-     */
-    protected $value;
+    #[ORM\Column(type: 'float', nullable: false)]
+    #[JMS\Expose]
+    protected ?float $value = null;
 
-    /**
-     * @ORM\Column(type="smallint", nullable=false)
-     * @JMS\Expose()
-     */
-    protected $pollutant;
+    #[ORM\Column(type: 'smallint', nullable: false)]
+    #[JMS\Expose]
+    protected ?int $pollutant = null;
 
-    public function getId(): int
+    #[ORM\Column(type: 'string', length: 32, nullable: true)]
+    #[JMS\Expose]
+    private ?string $tag = null;
+
+    public function setId(int $id): Data
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    public function getId(): ?int
     {
         return $this->id;
     }
 
-    public function getStation(): Station
+    public function getStation(): ?Station
     {
         return $this->station;
     }
@@ -77,19 +83,19 @@ class Data
         return $this->dateTime->format('Y-m-d H:i:s');
     }
 
-    public function getDateTime(): \DateTime
+    public function getDateTime(): \DateTimeInterface
     {
         return $this->dateTime;
     }
 
-    public function setDateTime(\DateTime $dateTime): Data
+    public function setDateTime(\DateTimeInterface $dateTime): Data
     {
         $this->dateTime = $dateTime;
 
         return $this;
     }
 
-    public function getValue(): float
+    public function getValue(): ?float
     {
         return $this->value;
     }
@@ -101,7 +107,7 @@ class Data
         return $this;
     }
 
-    public function getPollutant(): int
+    public function getPollutant(): ?int
     {
         return $this->pollutant;
     }
@@ -113,11 +119,28 @@ class Data
         return $this;
     }
 
+    public function getProvider(): ?string
+    {
+        return $this->station->getProvider();
+    }
+
     public function isIndexable(): bool
     {
         $dateTime = new \DateTimeImmutable();
         $dateTime->sub(new \DateInterval('P1W'));
 
         return $dateTime >= $this->dateTime;
+    }
+
+    public function getTag(): ?string
+    {
+        return $this->tag;
+    }
+
+    public function setTag(?string $tag): self
+    {
+        $this->tag = $tag;
+
+        return $this;
     }
 }
