@@ -5,17 +5,13 @@ namespace App\EventSubscriber;
 use App\SeoPage\SeoPageInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
+use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 
 class KernelEventSubscriber implements EventSubscriberInterface
 {
-    /** @var SeoPageInterface $seoPage */
-    protected $seoPage;
-
-    public function __construct(SeoPageInterface $seoPage)
+    public function __construct(protected SeoPageInterface $seoPage)
     {
-        $this->seoPage = $seoPage;
     }
 
     public static function getSubscribedEvents(): array
@@ -25,9 +21,9 @@ class KernelEventSubscriber implements EventSubscriberInterface
         ];
     }
 
-    public function onController(FilterControllerEvent $filterControllerEvent): void
+    public function onController(ControllerEvent $filterControllerEvent): void
     {
-        if (!$filterControllerEvent->isMasterRequest()) {
+        if (!$filterControllerEvent->isMainRequest()) {
             return;
         }
 
