@@ -2,51 +2,27 @@
 
 namespace App\Pollution\PollutionDataFactory;
 
+use App\Air\ViewModelFactory\MeasurementViewModelFactoryInterface;
 use App\Entity\Station;
-use App\Pollution\Box\Box;
-use App\Pollution\BoxDecorator\BoxDecoratorInterface;
 use App\Pollution\DataList\DataList;
 use App\Pollution\DataRetriever\DataRetrieverInterface;
 use App\Pollution\PollutantFactoryStrategy\PollutantFactoryStrategyInterface;
-use App\Pollution\StationFinder\StationFinderInterface;
-use Caldera\GeoBasic\Coord\CoordInterface;
+use Caldera\GeoBasic\Coordinate\CoordinateInterface;
 
 abstract class AbstractPollutionDataFactory implements PollutionDataFactoryInterface
 {
-    /** @var CoordInterface $coord */
-    protected $coord;
+    protected CoordinateInterface $coord;
 
-    /** @var StationFinderInterface $stationFinder */
-    protected $stationFinder;
+    protected DataList $dataList;
 
-    /** @var StationFinderInterface $boxDecorator */
-    protected $boxDecorator;
-
-    /** @var DataList $dataList */
-    protected $dataList;
-
-    /** @var DataRetrieverInterface */
-    protected $dataRetriever;
-
-    /** @var array $stationList */
-    protected $stationList = [];
-
-    /** @var PollutantFactoryStrategyInterface $strategy */
-    protected $strategy;
-
-    public function __construct(StationFinderInterface $stationFinder, BoxDecoratorInterface $boxDecorator, DataRetrieverInterface $dataRetriever, PollutantFactoryStrategyInterface $strategy)
+    public function __construct(protected MeasurementViewModelFactoryInterface $measurementViewModelFactory, protected DataRetrieverInterface $dataRetriever, protected PollutantFactoryStrategyInterface $strategy)
     {
-        $this->stationFinder = $stationFinder;
         $this->dataList = new DataList();
-        $this->boxDecorator = $boxDecorator;
-        $this->dataRetriever = $dataRetriever;
-        $this->strategy = $strategy;
     }
 
-    public function setCoord(CoordInterface $coord): PollutionDataFactoryInterface
+    public function setCoord(CoordinateInterface $coord): PollutionDataFactoryInterface
     {
         $this->coord = $coord;
-        $this->stationList = $this->stationFinder->setCoord($this->coord)->findNearestStations();
 
         return $this;
     }
@@ -54,7 +30,6 @@ abstract class AbstractPollutionDataFactory implements PollutionDataFactoryInter
     public function setStation(Station $station): PollutionDataFactoryInterface
     {
         $this->coord = $station;
-        $this->stationList = [$station];
 
         return $this;
     }
