@@ -2,13 +2,19 @@
 
 namespace App\Provider;
 
+use App\Air\Measurement\MeasurementInterface;
+
 abstract class AbstractProvider implements ProviderInterface
 {
-    /** @var StationLoaderInterface $stationLoader */
-    protected $stationLoader;
+    protected StationLoaderInterface $stationLoader;
 
     public function getStationLoader(): StationLoaderInterface
     {
         return $this->stationLoader;
+    }
+
+    public function providesMeasurement(MeasurementInterface $measurement): bool
+    {
+        return in_array($measurement::class, $this->providedMeasurements());
     }
 }
