@@ -2,8 +2,7 @@
 
 namespace App\Provider;
 
-use App\Entity\Station;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Doctrine\Persistence\ManagerRegistry;
 
 abstract class AbstractStationLoader implements StationLoaderInterface
 {
@@ -13,12 +12,8 @@ abstract class AbstractStationLoader implements StationLoaderInterface
     /** @var array $newStationList */
     protected $newStationList = [];
 
-    /** @var RegistryInterface $registry */
-    protected $registry;
-
-    public function __construct(RegistryInterface $registry)
+    public function __construct(protected ManagerRegistry $registry)
     {
-        $this->registry = $registry;
     }
 
     public function getNewStationList(): array
