@@ -4,43 +4,47 @@ namespace App\Controller;
 
 use App\Entity\City;
 use App\Entity\Station;
-use App\Entity\TwitterSchedule;
 use App\Pollution\PollutionDataFactory\PollutionDataFactory;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\Request;
+use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController as FrameworkAbstractController;
 
-abstract class AbstractController extends Controller
+abstract class AbstractController extends FrameworkAbstractController
 {
+    public function __construct(protected ManagerRegistry $managerRegistry)
+    {
+
+    }
+
+    /**
+     * @deprecated
+     */
+    public function getDoctrine(): ManagerRegistry
+    {
+        return $this->managerRegistry;
+    }
+
     protected function getStationListForCity(City $city): array
     {
         return $this->getDoctrine()->getRepository(Station::class)->findActiveStationsForCity($city);
     }
 
-    protected function createBoxListForStationList(PollutionDataFactory $pollutionDataFactory, array $stationList): array
+    protected function createViewModelListForStationList(PollutionDataFactory $pollutionDataFactory, array $stationList): array
     {
-        $stationsBoxList = [];
+        $stationViewModelList = [];
 
         /** @var Station $station */
         foreach ($stationList as $station) {
-            $stationsBoxList[$station->getStationCode()] = $pollutionDataFactory
+            $stationViewModelList[$station->getStationCode()] = $pollutionDataFactory
                 ->setStation($station)
                 ->createDecoratedPollutantList();
         }
 
-        return $stationsBoxList;
+        return $stationViewModelList;
     }
 
     protected function getCityBySlug(string $citySlug): ?City
     {
         return $this->getDoctrine()->getRepository(City::class)->findOneBySlug($citySlug);
-    }
-
-    protected function getTwitterScheduleByRequest(Request $request): ?TwitterSchedule
-    {
-        $scheduleId = $request->query->getInt('scheduleId');
-        $schedule = $this->getDoctrine()->getRepository(TwitterSchedule::class)->find($scheduleId);
-
-        return $schedule;
     }
 
     protected function findCityForName(string $cityName): ?City
