@@ -2,8 +2,7 @@
 
 namespace App\DependencyInjection\Compiler;
 
-use App\AirQuality\Calculator\AirQualityCalculatorInterface;
-use App\Command\StationCommand;
+use App\Provider\ProviderListInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\Reference;
@@ -12,11 +11,11 @@ class ProviderCompilerPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container)
     {
-        if (!$container->has(StationCommand::class)) {
+        if (!$container->has(ProviderListInterface::class)) {
             return;
         }
 
-        $stationCommand = $container->findDefinition(StationCommand::class);
+        $stationCommand = $container->findDefinition(ProviderListInterface::class);
 
         $taggedServices = $container->findTaggedServiceIds('air_provider');
 
